requirements.txt
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
coverage==6.3.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.4
django-filter==21.1
django-import-export==2.8.0
django-redis==5.2.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
executing==0.8.3
Faker==13.3.4
ipdb==0.13.9
ipython==8.2.0
jedi==0.18.1
MarkupPy==1.14
matplotlib-inline==0.1.3
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
redis==4.2.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tablib==3.2.1
toml==0.10.2
traitlets==5.1.1
wcwidth==0.2.5
wrapt==1.14.0
xlrd==2.0.1
xlwt==1.3.0
